Project 2 - Memory

+++ Project 2 - Memory/xv6/kernel/sysproc.c
@@ -102,5 +102,5 @@ sys_shmem_count(void)
 {
   int page_number;
   if (argint(0, &page_number) < 0) return -1;
-  return shmem_count(page_number);
+  return (int)shmem_count(page_number);
 }
